lots of improvements.  a tcp nfs client can now do an initial mount & df (which will return bogus information.
make nfssrv read nfs messages.  responds to most with error for
now.  the basic properties of incoming nfs requests are printed,
for debugging.

split reading & parsing of rpc messages.  rpc on top of tcp puts
in 4 extra bytes (length of fragment) in front of data.  udp doesn't.
so reading a message from tcp & udp is different.  once the buffer
is in, the parsing of the message is the same.  writing has changed
too:  for udp, the 52 byte destination has to be written.  so write
takes a "prefix" buffer to write to the file descriptor in front
of the packed response.

portmapper now listens on udp too.  my system doesn't even seem to
try tcp to portmapper, so tcp-only was no good.  also put in some
debug prints in the portmapper function handlers.

put a string in the "Badrpc" exception, which is printed as debug
output.  helps in debugging.  the parsing functions give more helpful
messages too (such as offsets in buffer where parsing failed).

fix the program numbers, they are 10000x instead of 1000x.

fix catching some exceptions.  there were some cases where the
compiler could've just warned...

implement rounding up to 4 bytes for each xdr data item.  packing/parsing
opaque/string was essentially broken before.