rewritten big parts of nfssrv.
much better handling of file handles.  before we generated file
handles consecutively and held them in our cache, until they were
evicted for other file handles.  turns out eg linux & openbsd nfs
clients don't re-lookup file handles when their current ones have
gone stale, so this gave problems.  we can now operate in two modes:
in the new mode file handles are always 8 bytes, holding the qid.
at startup, the user must give two directories, for in which you
can walk to and open/read/write/etc files by qid.path (as name),
and another one similar to it but where a walk to a qid.path walks
to its parent (for resolving ".." lookups).  the other mode has
changes too:  file handles can be up to 64 bytes in length.  if a
path is <= 63 bytes (as utf-8), we use the path with byte 16rff as
the file handle.  for larger paths we consecutively assign file
handles and evict them (nothing to do about that).  this makes us
really stateless most of the time.

we handle readdir & readdirplus much better.  we use a dir's mtime
& qid.vers as verifier.  the cookie is the offset in Dir's (starting
at 1).  this allows us to statelessly re-walk the directory and
continue a readdir(plus).  we properly evict the directory handles
after 90 seconds of inactivity.

we also cache opened files now.  before, we would open files for
each operation.  now we keep opened plain files (and their open
mode, OREAD or OWRITE) around for up to 15 seconds after last use.

we now optionally read a user & group file (e.g. /etc/passwd and
/etc/group) and use the uid/gids & users/groups for translating
between styx Dir's and nfs Attr's.  note that this not enforce
permissions of the connecting user, we still operate with the
permissions of the user that started nfssrv.