nfssrv: serve different nfs uid's by different progs that have their own mounted file system.
this means the different users properly get errors when they are
not allowed to do the operations they are attempting.  before this
(and still without -t on the command-line), all operations would
execute as the user that started nfssrv.

also update the manual page after this and other recent updates.